requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
faiss-cpu==1.7.4
filelock==3.25.2
Flask==3.1.3
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
openai==2.28.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==2.7.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.2.2
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
